fixed a bug with the progress bar by moving it into the webplayback file
